629369-null check in delegate conversion
We insert a null check into the code generated for a delegate
conversion, so that Nothing is converted to Nothing rather
that throwing a null exception.

Note that we continue to generate the incorrect null-throwing
code for conversions appearing in expression trees. We do that
because we don't want to break backward compatibility
(there is no way to translate into the null-check version). (changeset 1348760)