Add telemetry about the Analyzers node.
This change captures a telemetry data point the first time the Analyzers node or a child analyzer node is expanded in Solution Explorer. The intent is to capture data on how many users interact with this part of Solution Explorer.

Note that for any given node we only capture a data point the first time it is expanded. Collapsing and re-opening the node will not cause further data points to be captured, nor will adding or removing analyzers.

The data point also captures data on the number of analyzers, and the number of diagnostics in an analyzer. (changeset 1351286)