DevDiv #1060879: Workaround for EnC/EE interaction issue
The issue is that methods can point at other methods for their custom debug info.  However, the pointer contains no version information so we were assuming that both methods were on the same version.  This is not a safe assumption - each method is updated individually by EnC.  The real fix is to avoid forwarding custom debug info in EnC scenarios.  The workaround (for Preview) is to handle a null result from GetMethodByVersion.  With the workaround in place, the custom debug info (mainly, imports) is lost, but the entries in the Locals window do not disappear (and we don't get a Watson for this known issue).

Unfortunately, we do not have the testing infrastructure we need to add a regression test for this scenario (tracked by DevDiv #1064003).  However, I have manually verified the scenario against locally-built Dev14CTP binaries.

Dev14CTP checkin approved by MattGe. (changeset 1354173)