EE: Support object id in VB Expression Compiler
Note this uses Tomas' proposed object id syntax "$[1-9][0-9]+" rather than the Dev12 EE syntax "ObjId[1-9][0-9]+".

If we decide to use the Dev12 syntax instead, the Expression Compiler will need to call DkmInspectionContext.GetClrObjectIdType() when binding to check if there is an object with the given id that may shadow another symbol in scope with the same name.

CR: acasey (changeset 1357770)