PERF: Reduce size of SyntaxFactory's static constructor.
The size (19KB in the case of VB) is because of the initialization code for the NodeTypes static field.
While this change doesn't change the total size of generated code, it does move it out of the static constructor into its own method. The change is based on the observation that NodeTypes is used only within the GetSerializationData method. The result is transient - it's concatenated with other object sequences. So there's no need to keep this large object array in a static field either. (changeset 1358397)