Fix for bug 1069573: Roslyn crashes due to "Found invalid data while decoding" even without any analyzers/code fix providers imported.
The crash happens when attempting LoadTextAsync for non-text additional documents in a project file. We throw an InvalidDataException for this scenario, bringing down VS. This fix guards against the crash by defaulting to an empty text document for non-text files. In future, when we add support for non-text additional documents, such that they are based on a non-text document state, this change will be reverted. (changeset 1361214)