Clarify a comment in SourceMethodSymbol.
One paragraph in the comment for SourceMethodSymbol.LazyMethodChecks was very confusing to me.  The commented out SpinWaitComplete code was simply incorrect given the fact that LazyMethodChecks is reentrant.  Instead, I clarified the reentrancy and explained why it would be incorrect to assert that when we exit the method via the else case, FinishMethodChecks is true. (changeset 1361245)