Fix for bug 1068099: Warning level for info/hidden diagnostics should be non-zero
Ensure that warning level for non-error diagnostics is non-zero. This is required for backward compat, as the only way to suppress warnings in C# used to be /warn:0.  If user has command line with /warn:0, they should continue to only see errors in their output. Prior to this change, we also show info messages in the output. (changeset 1362306)