Bugfix 1068969 - ArgumentOutofRangeException after renaming the namespace in controllers
When using Rename Tracking to rename "Abc" to "Abcdef", the RenameTrackingCommitter first rolls back the locally changed identifier, in this case back to "Abc", before calling RenameSymbol. For linked files, the RenameTrackingCommitter was only updating this identifier in the current context document which caused an inconsistent solution to be passed into the Rename API. We now roll back the identifier changes in all linked documents before calling RenameSymbol.

Additionally, the underlying LinkedFileReferenceFinder now defends against inconsistent APIs in the event one is passed in. (changeset 1364246)