DevDiv #913022: Emitting extern alias custom debug info
When the target of a using alias used an extern alias as a qualifier, we used to emit the first alias on the containing assembly.  This was problematic since that extern alias might not be in scope.  Now we take into account which extern aliases are in scope.  This is still not as nice as Dev12's behavior - matching the alias in source - but it's cheaper and should be equivalent.

CR: ToMat; AnGocke (changeset 1367143)