When mapping a symbol through a metadata reference, ignore assembly names.
This is to deal with a fairly messy situation around type forwarding. Consider that you have a project, referencing System.Runtime, that has a method that takes an Int32 as a parameter. That parameter, per the symbols, comes from System.Runtime. If you have a non-portable project reference that uses that method from metadata, it's view is the method has a paremeter of type Int32 from mscorlib. When we resolve method signatures, the code currently assumes that you can simply create a SymbolKey for both sides and that they must be equal (including assembly name) across these two projects.

Sadly, we can't just change how SymbolKeys get created. I was hoping that for the non-portable project with the facades we could simply use the facade as the "real" name so then it would match. However, portable projects also have an mscorlib that forwards to System.Runtime, and so things still wouldn't match. So my fix is just to ignore assembly containment when mapping between compilations.

Fundamentally, SymbolKey is being abused for two different things: resolving symbols between compilations (of potentially different frameworks) and also for mapping symbols between different versions of the same logical project. If we avoided using SymbolKey to implement SymbolFinder.FindSourceDefinitionWorkerAsync we could instead do something smarter here looking at how types forward, but as it stands that'd require lots of work for unclear benefit. (changeset 1368497)