DevDiv #1066489: Debugging inside types with names containing dot
The problem is that the mangled name for an static machine type looks like
"<" + methodName + ">d__" + uniqueId.  However, methodName will have dots in it for explicit interface implementations (e.g. "<I.F>d__0").  Unfortunately, the native compiler emits such names in a very strange way: everything before the last dot goes in the namespace (!!) field of the typedef.  Since state machine types are always nested types and since nested types never have explicit namespaces (since they are in the same namespaces as their containing types), it should be safe to check for a non-empty namespace name on a nested type and prepend the namespace name and a dot to the type name.  After that, debugging support falls out.

CR: AnGocke; ToMat; AlekseyT (changeset 1368701)