Remove LanguageVersion.Experimental
The principal change here is to remove the language version "experimenal." However, there are a number of places in the code where we either set the language version to experimental or check whether experimental features are enabled. Those are largely vestigal at the moment as we have no experimental features. However as soon as we add experimental features that code will once again be useful. Rather than deleting that code now, I wanted to leave it in and make it easy to find and update as features are added. So I'm adding a stubbed implementation of the API proposed for future use for enabling experimental features.

After Dev14 our plan is to have experimental features enabled one-by-one rather than all under a single umbrella "experimental" flag. Since language features affect the parser, this is part of the parse options. This changeset adds a stubbed version of the proposed APIs:

public class ParseOptions
{
    /// <summary>
    /// Enable some experimental language features for testing.
    /// </summary>
    public ParseOptions WithFeatures(IEnumerable<KeyValuePair<string, string>> features);

    /// <summary>
    /// Returns the experimental features.
    /// </summary>
    public abstract IReadOnlyDictionary<string, string> Features
    {
        get;
    }
} (changeset 1369555)