While looking through some code last week I noticed that several types had fields of type ImmutableArray<T> but took IEnumerable<T> parameters to the constructors. Internally it would convert to immutable array via null checks and type tests. The general pattern was
class Widget {
 ImmutableArray<Bolt> bolts;

 Widget(IEnumerable<Bolt> b) {
 bolts = b == null ? ImmutableArray<Bolt>.Empty : b.ToImmutableArray();
 }
}

This seems like the wrong approach to immutable construction. The best place to make the immutable construction is in the caller because it can take advantage of types like ImmutableArray<T>.Builder instead of wasting a List<T>. Also in the ideal case where the caller has an ImmutableArray<T> in hand, the ideal scenario, it avoids a boxing allocation.

To fix I moved a few of these APIs to take ImmutableArray<T> as a parameter and followed the string out to the calling methods.

If this change is generally acceptable there are other scenarios where it should be applied. Most notably CodeFixContext.Diagnostics. The diagnostics member is constantly transitioning between IEnumerable<T> and ImmutableArray<T>. It would save a lot of type tests and allocations to just make it ImmutableArray<T>.
 (changeset 1370116)