Fix subtle bug in SymbolTreeInfo that HasSymbols might lie.
The underlying data structure stored in SymbolTreeInfo is a list of symbols sorted by name. Since we need to query it both case sensitively and case insensitively, we sort the list insensitively. Finding elements is done simply by binary searching insensitively, finding an element that matches the name insensitively, going forwards/backwards to collect all the elements that match insensitively, and filtering to sensitive matches if necessary.

There was a typo in the "walk back" code: if doing a case sensitive search and the binary search happened to give a case insensitive match, and all the correct sensitive matches were before the place the binary search found, it would not find the case sensitive match.

I attempted to (and failed) to write a reasonable unit test for this. It seems for the same assembly metadata the sort order might be different since the sort isn't necessarily stable. I could write a unit test that set up this specific list of nodes that made this fail, but it's such a specific unit test it wouldn't likely catch future mistakes. (changeset 1371484)