Fix struct layout calculation for structs without [StructLayoutAttribute].
The calculation of struct layout for structs without [StructLayoutAttribute] was incorrect with respect to Dev11. Where Dev11 emitted [StructLayout(LayoutKind.Sequential, Size = 1)] for structs without any instance fields--synthesized or otherwise--we were emitting the same only for structs that had fields (including static fields) explicitly defined in source. This fixes the calculation to agree with Dev11.
***NO_CI***
 (changeset 1371608)