Invalid IL generated for use of event assignment expression
Event assignment in .winmdobj is a call to a void method, so the assignment expression is now treated as a void expression rather than an expression with the value of the RHS. This matches the behavior of the native compiler.

We could change the lowering of E = expr in the future to { tmp = expr; E = tmp; tmp } if there's sufficient value in doing so. (changeset 1374463)