Fix up SymbolDisplay for extension methods
SymbolDisplay has a SymdolDisplayExtensionMethodStyle which is either Static or Instance and the default today is Static. However the implementation of symboldisplay is such that if you pass in a ReducedExtensionMethodSymbol an ask for Static style, it will still give you instance style. In other words, it doesn't force static style. Having the method kind determine the signature (and not forcing one way or another) is useful to where you want to get the string based on context like quickinfo at a callsite of an ext method versus at definition.

So I'm switching the StaticMethod style to call it Default which then retains existing behavior and is not a breaking change. I'm adding a new StaticMethod style which will force static style. This is needed because the expandreduce engine calls symboldisplay and compares strings to make sure that the symbol didn't change on expand\reduce and needs to force the style to static or instance. Today it forces it tries to static which doesnt really force it.
 (changeset 1375480)