Break the dependency between the Microsoft.CodeAnalysis.EditorFeatures.Text package and the Microsoft.CodeAnalysis.Features package.
We don't want to ship the Features package, because it has no public surface area, but we do need to ship the Text package.  Unfortunately, the implementation of the Text package currently depends on the internal PrimaryWorkspace type which lives in Features.  Move it to the Workspaces layer instead. (changeset 1375888)