Fix Watson crashes in 1024401.
At least some of the cab files indicate that the compiler is crashing while emitting custom attributes. This appears to be due to the compiler-generated DebuggerDisplayAttribute which is emitted on anonymous types. The current synthesis of the attribute proceeds without checking if a properly formed DebuggerDisplayAttribute type is present in the compilation.

This changes the attribute synthesis helpers to check if the well-known members exist before synthesizing the attribute. (changeset 1376039)