Never omit conditional calls in the debugger (VB)
The VB expression compiler doesn't directly invoke the LocalRewriter, so we had to find a different way to pass through the flag.  The PEModuleBuilder seemed the least obtrusive.

Also flipped the meaning of the (existing) flag in C# to match the way it is used (and VB).

CR: VSadov; AlekseyT; ChuckS (changeset 1376791)