DevDiv #1022632: Managed address-of
Thanks to Tomas' suggestion that we use IntPtr in place of reference types and Vlad's excellent code generator, this feature basically only requires the suppression of a couple diagnostics.

There is, however, a small change (improvement?) from dev12.  Suppose "str" is a string.  Then dev12 types "&str" as "string&*", which is probably the most accurate .NET description of the type.  However, Roslyn does not support & types, so it uses the representation "IntPtr*".  The value of the pointer is the same, and dereferencing yields the same pointer, albeit as an IntPtr, rather than as a string&.  This might actually be an improvement, since it is clear what operations can be performed on an IntPtr, whereas a string& can only be manipulated through trial and error (e.g. it turns out that it supports ==).

I think it would be reasonable to add a VS integration test for this scenario, but I don't know of a good way to extract the pointer value so that it can be re-used with the @ syntax, which can actually be verified.  Also, our base drop of VS can't launch the Memory window, which it would also be nice to verify.

CR: VSadov; ChuckS (changeset 1377887)