Fix 1032909: Spurious warning after analyzer load errors
When we crack open an analyzer dll and try to instantiate the analyzers, we report warnings if we cannot instantiate them. If we are unable to instantitate any analyzer in that assembly, we report that there are no analyzers as well which is silly. Fixing it by tracking when we report an error and not issuing the noanalyzers warning in that case. (changeset 1378514)