Clean up unnecessary StyleCop inclusion and fix packages.config.
We had a number of projects that included the StyleCop NuGet package only to also disable StyleCop. We also had some projects that were referencing assemblies out of packages but didn't say so in packages.config, or had stuff in packages.config not reflected in the actual project file. (changeset 1378738)