Track whether or not field accesses must be values rather than variables.
Certain field accesses must always be values rather than variables in order to properly implement uses of fields of readonly variables (e.g. foreach iteration variables). This change adds a field to track whether or not a given field access must always be a value and adjusts code generation appropriately: taking the address of such an access instead takes the address of a temporary copy of the field's value. This is necessary in particular for method calls on by-value accesses.
***NO_CI***
 (changeset 1378889)