Emitting synthetic ctor and cctor in the same relative order as in Dev12.
Changing order causes unnecessary noise in metadata diffing tools. Also we have seen some evidence of existing code taking unintentional dependency on the order of constructors as returned via reflection. That is arguably a bug since the order is not specified nor guaranteed. On the other hand we do not have any serious reasons for moving that cheese, so we will try keeping the emit order of synthetic ctors as it was.

Dev12 C# compiler appears to emit synthetic constructors after all other methods - first ctor, and then cctor, if needed.
 (changeset 1379237)