Add a nop after a StopStatement in VB
When a method like System.Diagnostics.Debugger.Break() is called, we emit a nop after a call, so the debugger can determine the current instruction pointer. However, StopStatement, which will be lowered into System.Diagnostics.Debugger.Break(), is marked as a compiler generated node, so this nop is not emitted, causing the debugging experience disrupted. This shelveset is not to mark the lowered node from a  StopStatement as a compiler generated, so in the later codegen phase, a nop would be emitted as in Dev12. (changeset 1379591)