Fix 1082030: Generate Property does not appear for property in VB Object Initializer
In VB, the assignments to object initialzers are not ExpressionSyntaxes but is a NamedFieldInitializerSyntax and this needs to be special cased for GFU. In C#, it's just an assignmentexpression and so everything works fine.

One relaed issue is that both in C# and VB, we started showing Generate Local code fix which doesn't make sense here. Fixing that by ensureing that we don't offer generate local when we are in a member context.
 (changeset 1381118)