Limit the size of the MemoryStream instances that we keep in the pool.
Unlimited size of cached memory streams causes memory retention problems in big projects.
The re is a bug report about a project that results in 84MB retained in this pool after the first compile.

Statstics of Roslyn self-compile shows that vast majority of the memory streams that we use are short.

0-512             120K
512-1024       4K
1024-1536     714
1536-2048     407
. . .
. . .
long tail follows with couple really gigantic streams at the end.

Also limited the number of pool items to 32. That limits overall size of this cache to 32KB
.
 (changeset 1381638)