Fix a crash that occurs when lowering await expressions with dynamically-typed results.
The await rewriter was attempting to manually replace "dynamic" with "object" in an expression's awaiter type and related methods for await expressions that are typed as "dynamic". Not only is this unnecessary--this sort of rewrite will be done at emit time--it is also incorrect given the pattern-based nature of await binding.

I also simplified the code that finds and validates the various methods and types needed for an await expression while making this fix.
***NO_CI***
 (changeset 1383986)