Apply the AppDomain's policy when loading analyzer dependencies.
When the AppDomain.CurrentDomain.AssemblyResolve event is fired to resolve an analyzer's dependency, the ResolveEventArgs.Name property has the name of the desired assembly *before* any policies (like binding redirects) have been applied. Since we're doing our own assembly resolution it is up to us to apply policy, and we're advised that this is generally the correct thing to do. (changeset 1383995)