Fix the expander to not crash on ?.
When only a simple name is allowed in a context the expander shouldn't try to expand anything. Teach the expander that in C#, a MemberBindingExpression(the thing that follows ?) can only have a simplename after it. In VB, this is just a memberaccess and so the expander already knows what to do. (changeset 1385689)