Ensure that GetSemanticInfo() returns the correct members for the LHS of a conditional access expression.
The syntax node for the LHS of a conditional access expression appears in two places:
- Inside its own BoundExpression, tracked by BoundConditionalAccess.Receiver
- Inside the BoundConditionalReceiver that is attached to the RHS of the conditional access expression, tracked by BoundConditionalAccess.AccessExpression

Or, diagramatically (let "SyntaxNode" be the syntax node for the LHS of the conditional access expression):

      BoundConditionalAccessExpression
      /                              \
BoundExpression                    .....
      |                              |
  SyntaxNode               BoundConditionalReceiver
                                     |
                                 SyntaxNode

This has the effect of confusing the semantic model when looking for symbol information for SyntaxNode: because the appearance of SyntaxNode in the RHS of the BoundConditionalAccessExpression is lower in the bound tree than the appearance of SyntaxNode in the LHS, the BoundConditionalReceiver associated with the RHS will be used to lookup symbols. The primary effect of this is a failure to return the actual method called in situations such as the following:

class C
{
    public int I;

    public static C M(long l) { return null; }
    public static C M(int i) { return null; }

    public static int Main()
    {
        return (int)(M(0)?.I);
    }
}

If semantic information is requested for "M" in the call to "M(0)", the returned information will contain both overloads of "M" instead of the actual overload chosen.

The actual fix is simple:
- Mark all BoundConditionalReceiver nodes as compiler generated, thus eliminating them from consideration for MemberSemanticModel.GetBoundNodes
- Stop digging through the syntax tree for conditional expressions when getting the bindable parent node of an expression

Unfortunately, the recommendation service took a dependency on seeing the BoundConditionalReceiver in the tree when requesting type information for the bound LHS of a member binding expression via SemanticModel.GetTypeInfo: if the type of the original receiver for the conditional access is T?, the type of the BoundConditionalReceiver is T. Instead of relying on this behavior, the recommendation service has been updated to perform this transformation manually.
***NO_CI***
 (changeset 1386679)