Bugfix 1006296 - "GTD should not automatically change project context"
Before navigating to a symbol defined in a given document, check whether a related linked or shared document is the active context document and use that document instead. For regular files (non-shared and non-linked) and closed linked files, this is always the provided document. For open linked files and open shared files, the active context is already tracked by the workspace and can be looked up directly. For closed shared files, the document in the shared project's SharedItemContextHierarchy is preferred. (changeset 1387121)