Bug fixes for index intializers.
Index expressions in the index initializers should be evaluated exactly once.
In a case of a nested initializer, we were evaluating as many times as we have members in the RHS. That resulted in a number of bugs reported.

Working on this fix revealed some more issues with params and named indices not being supported properly.

Fixes #449
 (changeset 1388102)