Conversions for interpolated strings v8
We also add all of the relevant optimizations requested for current and future
formatting performance improvements. We now use full overload resolution
(using a new binder type specialized for synthesizing invocations) to
select an appropriate string.Format method or format factory method, which
enables it to handle params, generics, arbitrary argument conversions, etc, and we
optimize the special case of no fill-ins by processing the {{ and }} escapes,
producing a literal in the generated code. (changeset 1388224)