Fix issues related to Microsoft.CodeAnalysis.Test.Resources.Proprietary.
We're supposed to be picking up this assembly from a NuGet package, but one project still has a project-to-project reference instead. This means different projects are building against different versions of the assembly, and this can cause the build to fail depending on the order in which projects are handled.

Also, a couple of projects had references to two different versions of the NuGet. At the very least the assembly would show up twice in Solution Explorer (once as a resolved reference, and once as an unresolved reference) but probably also caused build issues on a system with both versions present in the packages directory. (changeset 1389738)